requirements.txt
# Face Recognition System Requirements
# Compatible with macOS and Linux

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning & Face Recognition
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
scipy==1.11.4
scikit-image==0.22.0
easydict==1.9

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
cryptography==41.0.7

# API & Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0

# Optional: Monitoring & Logging
# prometheus-client==0.19.0
# sentry-sdk==1.38.0
